requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.54.3
playwright==1.40.0
aiohttp==3.9.1
requests==2.31.0
pypdf2==3.0.1
pandas==2.1.4
numpy==1.26.2
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
httpx==0.27.0
